requirements.txt
requests>=2.31.0
rich>=13.7.0
python-dateutil>=2.8.2
